requirements.txt
fastapi
uvicorn
gunicorn
requests
python-jose[cryptography]
pytest
pytest-cov
prometheus-fastapi-instrumentator
httpx